feat: change to gRPC (#11)
BREAKING CHANGE: This removes the normal http
call to `/ca` and `/csr` in favor of a gRPC api.
Head over to `proto/pki.proto` to see the required
gRPC calls.

BREAKING CHANGE: This further improves the configuration possibility of the PKI
by using env and command line arguments. See
details in the readme.